# swagger/swagger.json
{"swagger":"2.0","info":{"title":"viron goa","description":"viron goa service","version":""},"consumes":["application/json"],"produces":["application/json"],"paths":{"/signin":{"post":{"tags":["viron"],"summary":"signin viron","description":"Creates a valid JWT","operationId":"viron#signin","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/SigninPayload"}}],"responses":{"204":{"description":"No Content","headers":{"Authorization":{"description":"Generated JWT","type":"string"}}},"401":{"description":"Unauthorized"},"500":{"description":"Internal Server Error"}}}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","externalDocs":{"description":"Swagger JSON","url":""},"operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"security":[{"jwt":[]}]}},"/viron":{"get":{"tags":["viron"],"summary":"get viron","description":"Get viron menu","operationId":"viron#get","produces":["application/vnd.vironsetting+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Vironsetting"}}},"security":[{"jwt":[]}]}},"/viron_authtype":{"get":{"tags":["viron"],"summary":"authtype viron","description":"Get viron authtype","operationId":"viron#authtype","produces":["application/vnd.vironauthtype+json; type=collection"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/VironauthtypeCollection"}}}}},"/{filepath}":{"get":{"summary":"Download viron-client/v1","externalDocs":{"description":"viron client","url":""},"operationId":"client#/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}}}}},"definitions":{"Query":{"title":"Mediatype identifier: application/vnd.query+json; view=default","type":"object","properties":{"key":{"type":"string","description":"key","example":"Id autem ratione dolor."},"type":{"type":"string","description":"type","example":"Aut et deserunt voluptates."}},"description":"viron query (default view)","example":{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},"required":["key","type"]},"SigninPayload":{"title":"SigninPayload","type":"object","properties":{"email":{"type":"string","description":"ID or Email","example":"identify key"},"password":{"type":"string","description":"Password","example":"zil3mqvnk2","maxLength":256}},"example":{"email":"identify key","password":"zil3mqvnk2"},"required":["email","password"]},"Vironapi":{"title":"Mediatype identifier: application/vnd.vironapi+json; view=default","type":"object","properties":{"method":{"type":"string","description":"name","example":"Est corrupti voluptatem."},"path":{"type":"string","description":"path","example":"Autem perferendis quia provident."}},"description":"viron api (default view)","example":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"required":["method","path"]},"Vironauthtype":{"title":"Mediatype identifier: application/vnd.vironauthtype+json; view=default","type":"object","properties":{"method":{"type":"string","description":"method","example":"POST"},"provider":{"type":"string","description":"auth provider","example":""},"type":{"type":"string","description":"auth type","example":"signin"},"url":{"type":"string","description":"url","example":"/signin"}},"description":"viron authtype media (default view)","example":{"method":"POST","provider":"","type":"signin","url":"/signin"},"required":["type","provider","url","method"]},"VironauthtypeCollection":{"title":"Mediatype identifier: application/vnd.vironauthtype+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Vironauthtype"},"description":"VironauthtypeCollection is the media type for an array of Vironauthtype (default view)","example":[{"method":"POST","provider":"","type":"signin","url":"/signin"},{"method":"POST","provider":"","type":"signin","url":"/signin"}]},"Vironcomponent":{"title":"Mediatype identifier: application/vnd.vironcomponent+json; view=default","type":"object","properties":{"api":{"$ref":"#/definitions/Vironapi"},"name":{"type":"string","description":"name","example":"Esse nemo omnis qui quis odio."},"pagination":{"type":"boolean","description":"pagination","example":true},"primary":{"type":"string","description":"primary key","example":"Vero a accusamus ad et."},"query":{"type":"array","items":{"$ref":"#/definitions/Query"},"description":"query","example":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}]},"style":{"type":"string","description":"style","example":"Nesciunt provident aut deleniti."},"table_labels":{"type":"array","items":{"type":"string","example":"Voluptates qui."},"description":"table label","example":["Voluptates qui."]}},"description":"viron component (default view)","example":{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},"required":["name","style","api"]},"Vironpage":{"title":"Mediatype identifier: application/vnd.vironpage+json; view=default","type":"object","properties":{"components":{"type":"array","items":{"$ref":"#/definitions/Vironcomponent"},"description":"pages","example":[{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]}]},"group":{"type":"string","description":"group","example":"At molestias necessitatibus quasi."},"id":{"type":"string","description":"id","example":"Iusto rerum similique provident."},"name":{"type":"string","description":"name","example":"Vitae dolorem."},"section":{"type":"string","description":"section","example":"Rerum sint molestiae."}},"description":"viron page (default view)","example":{"components":[{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]}],"group":"At molestias necessitatibus quasi.","id":"Iusto rerum similique provident.","name":"Vitae dolorem.","section":"Rerum sint molestiae."},"required":["id","name","section","group","components"]},"Vironsetting":{"title":"Mediatype identifier: application/vnd.vironsetting+json; view=default","type":"object","properties":{"color":{"type":"string","description":"color","example":"Ut quibusdam delectus."},"name":{"type":"string","description":"name","example":"Expedita hic eum magni placeat pariatur."},"pages":{"type":"array","items":{"$ref":"#/definitions/Vironpage"},"description":"pages","example":[{"components":[{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]}],"group":"At molestias necessitatibus quasi.","id":"Iusto rerum similique provident.","name":"Vitae dolorem.","section":"Rerum sint molestiae."}]},"tags":{"type":"array","items":{"type":"string","example":"Quo omnis ut quia consequuntur qui."},"description":"tags","example":["Quo omnis ut quia consequuntur qui.","Quo omnis ut quia consequuntur qui."]},"theme":{"type":"string","description":"theme","example":"A consequatur distinctio possimus in dolor repellendus."},"thumbnail":{"type":"string","description":"thumbnail","example":"Quos quia."}},"description":"viron setting (default view)","example":{"color":"Ut quibusdam delectus.","name":"Expedita hic eum magni placeat pariatur.","pages":[{"components":[{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]},{"api":{"method":"Est corrupti voluptatem.","path":"Autem perferendis quia provident."},"name":"Esse nemo omnis qui quis odio.","pagination":true,"primary":"Vero a accusamus ad et.","query":[{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."},{"key":"Id autem ratione dolor.","type":"Aut et deserunt voluptates."}],"style":"Nesciunt provident aut deleniti.","table_labels":["Voluptates qui."]}],"group":"At molestias necessitatibus quasi.","id":"Iusto rerum similique provident.","name":"Vitae dolorem.","section":"Rerum sint molestiae."}],"tags":["Quo omnis ut quia consequuntur qui.","Quo omnis ut quia consequuntur qui."],"theme":"A consequatur distinctio possimus in dolor repellendus.","thumbnail":"Quos quia."},"required":["name","color","theme","pages","tags","thumbnail"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"InternalServerError":{"description":"Internal Server Error"},"Unauthorized":{"description":"Unauthorized"}},"securityDefinitions":{"jwt":{"type":"apiKey","description":"\n\n**Security Scopes**:\n  * `api:access`: API access","name":"Authorization","in":"header"}}}